Java packages refactoring
