+++ app/src/main/java/com/example/lab4_multimedia/media_player/PlaylistControlsFragment.java
@@ -1,4 +1,4 @@
-package com.example.lab4_multimedia;
+package com.example.lab4_multimedia.media_player;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -9,6 +9,8 @@ import androidx.annotation.NonNull;
 import androidx.constraintlayout.widget.ConstraintLayout;
 import androidx.fragment.app.Fragment;
 import androidx.fragment.app.FragmentResultListener;
+
+import com.example.lab4_multimedia.R;
 
 public class PlaylistControlsFragment extends Fragment {
     private ConstraintLayout previous_layout;
