+++ app/src/main/java/com/example/lab4_multimedia/media_player/SongInfoFragment.java
@@ -1,4 +1,4 @@
-package com.example.lab4_multimedia;
+package com.example.lab4_multimedia.media_player;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -10,6 +10,8 @@ import android.widget.TextView;
 import androidx.annotation.NonNull;
 import androidx.fragment.app.Fragment;
 import androidx.fragment.app.FragmentResultListener;
+
+import com.example.lab4_multimedia.R;
 
 public class SongInfoFragment extends Fragment {
     public enum RegisterFor {
