+++ app/src/main/res/layout/activity_media_player_main.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MediaPlayerMainActivity">
+    tools:context=".media_player.MediaPlayerMainActivity">
 
     <androidx.fragment.app.FragmentContainerView
         android:id="@+id/playlist_control_container"
