+++ app/src/main/res/layout/activity_sign_in.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SignInActivity">
+    tools:context=".onboarding.SignInActivity">
 
     <com.google.android.material.appbar.MaterialToolbar
         android:id="@+id/sign_in_navigation_bar"
