+++ app/src/main/res/layout/fragment_cloud_media_explorer.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".CloudMediaExplorerFragment">
+    tools:context=".cloud_media_explorer.CloudMediaExplorerFragment">
 
     <TextView
         android:id="@+id/cloud_library_title"
