+++ app/src/main/res/layout/fragment_media_player.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".MediaPlayerFragment">
+    tools:context=".media_player.MediaPlayerFragment">
 
     <com.google.android.exoplayer2.ui.PlayerControlView
         android:id="@+id/player_view"
