+++ app/src/main/res/layout/fragment_playlist_controls.xml
@@ -4,7 +4,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".PlaylistControlsFragment">
+    tools:context=".media_player.PlaylistControlsFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/playlist_control_previous_layout"
