+++ app/src/main/res/layout/fragment_song_info.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".SongInfoFragment">
+    tools:context=".media_player.SongInfoFragment">
 
     <androidx.constraintlayout.widget.Guideline
         android:id="@+id/song_info_cover_left_guideline"
